+++ src/App.vue
@@ -17,6 +17,7 @@ import useLogin from '@/context/login';
 import DialogModal from "@/components/layout/DialogModal.vue";
 import { l } from '@/locales';
 import { router } from './router';
+import LanguagePopup from "@/components/settings/LanguagePopup.vue";
 
 
 const login = useLogin();
@@ -24,10 +25,12 @@ const state = reactive<{
     showInactivityPopup: boolean
     showInactivityWarningPopup: boolean
     logoutCountdown: number
+    showLanguagePopup: boolean
 }>({
     showInactivityPopup: false,
     showInactivityWarningPopup: false,
     logoutCountdown: 0,
+    showLanguagePopup: false,
 });
 let logoutCountdownInterval: number | undefined;
 
@@ -70,6 +73,9 @@ onMounted(() => {
         (user) => {
             if (user && !import.meta.env.VITE_DEBUG_DISABLE_SESSION_TIMEOUT) {
                 startInactivityListener(onInactivityTimeout, setInactivityWarningPopupVisibility);
+                // Check if language has been selected
+                const hasSelectedLanguage = localStorage.getItem("hasSelectedLanguage") === "true";
+                state.showLanguagePopup = !hasSelectedLanguage;
             }
             else {
                 stopInactivityListener();
@@ -88,8 +94,8 @@ onMounted(() => {
 </script>
 
 <template>
+    <LanguagePopup v-if="state.showLanguagePopup" />
     <DialogModal :active="state.showInactivityPopup"
-        @focus-first-button="(btn: HTMLElement | null) => btn?.focus()"
         :buttons="[{ loc_key: l.button_accept, onClick: closeInactivityPopup }]">
         {{ $t(l.logout_inactivity_message) }}
     </DialogModal>