+++ src/context/chat.ts
@@ -317,6 +317,24 @@ async function streamResponse() {
     } else {
         console.warn("AI service is unavailable");
     }
+    // Generate a random delay between between 1000 ms and 500 ms
+    const randomDelay = Math.floor(Math.random() * (1000 - 500 + 1)) + 500;
+
+    // Create a delay promise that resolves after the random delay
+    const delayPromise = new Promise<void>((resolve) => {
+        setTimeout(() => {
+            if (AireServices.AI) {
+                useChatbot().makeBusy();
+        
+                const input = getChatbotInputData()
+                AireServices.AI.stream(input, receiver, errorHandler);
+            } else {
+                console.warn("AI service is unavailable");
+            }
+            console.debug(`Delay of ${randomDelay} ms complete.`);
+            resolve();
+        }, randomDelay);
+    });
 }
 
 // Return true if event handled