+++ docs/development/CHANGELOG.md
@@ -11,10 +11,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/), and this 
 ### Added
 - More keyboard navigatable elements/parts and some changes to UI behaviour when navigating.
 
+## [1.0.2] - 2024-10-31
+### Fixed
+Localizations changed for all languages realted to tooltipa, and some others.
+
 ## [1.0.1] - 2024-10-27
 ### Fixed
 - Minor UI alignment issues on the setttings panel for a layout smaller than a normal desktop 1920x1080p.
 
+### Added
+- created a new Popup menu to choose UI language for user.
+- Generate a random delay between between 1000ms and 500ms for bot to respond to make the bot more human.
 
 ## [1.0.0] - 2024-05-21
 ### Added